v2.4.9.4 -> v2.4.9.5
  - Merge with Alan
  - Trond Myklebust: NFS fixes - kmap and root inode special case
  - Al Viro: more superblock cleanups, inode leak in rd.c, minix
  directories in page cache
  - Paul Mackerras: clean up rubbish from sl82c105.c
  - Neil Brown: md/raid cleanups, NFS filehandles
  - Johannes Erdfelt: USB update (usb-2.0 support, visor fix, Clie fix,
  pl2303 driver update)
  - David Miller: sparc and net update
  - Eric Biederman: simplify and correct bootdata allocation - don't
  overwrite ramdisks
  - Tim Waugh: support multiple SuperIO devices, parport doc updates

+++ drivers/macintosh/nvram.c
@@ -18,7 +18,7 @@
 
 #define NVRAM_SIZE	8192
 
-static long long nvram_llseek(struct file *file, loff_t offset, int origin)
+static loff_t nvram_llseek(struct file *file, loff_t offset, int origin)
 {
 	switch (origin) {
 	case 1:
